requirements.txt
certifi==2019.11.28
flake8==3.7.9
ipdb==0.12.3
mock==3.0.5
slackclient==2.4.0
yapf==0.29.0
